Remount btrfs with nobarrier before suspend
In some cases, btrfs might still have data remaining to be flushed to disk
before we suspend. This would mean the suspend script will hang after
suspending the LUKS device, so the user would have to manually unlock the
device or reboot the computer. This is probably the same problem with ext4 (and
possibly other filesystems!)