Fix incorrect exit code check in suspend script
I split a check for exit codes 0 or 1 into two statements:

    [[ $? == 0 ]] && break
    [[ $? == 1 ]] && break

If cryptsetup luksResume returns with exit code 2 (no key found i.e. invalid
passphrase), then the first statement will return exit code 1, which would
result in the suspend script in continuing.

This is incorrect: we want 0 (normal, unlock success) or 1 (already unlocked)
from cryptsetup, but the above causes the script to just continue which leads
to a hang. Oops.